Update main_TLS_calib.m
Initializing the structure
A filter would also be good in the structure
Why do use a loop if you only iterate it trough one. Then it would be better to use a for loop if you iterate through it a fixed number of times.
I think the lines: 130, 132, 133, 134, 136, 137 are unnecessary --> those value do not change
Is the loop starting at line 122 for detecting gross outliers (like presorting to avoid false starting point) or is it actual adjustment. If it is for the adjustment, we need to update the P matrix, the x values and scans in spherical coordinates.