Update RunGMMAdjust.m
I would not pack all the data from the structure into separate variables. Just call the structure.variable.
Then we should introduce a threshold like if the changes made to the x_0 are undeneath a certain level and then giving him max number of iterations to prevent an endless calculating.